+++ crates/symbolicator/src/services/objects/data_cache.rs
@@ -21,6 +21,7 @@ use symbolic::debuginfo::{Archive, Object};
 use tempfile::tempfile_in;
 
 use crate::cache::{CacheKey, CacheStatus};
+use crate::logging::LogError;
 use crate::services::cacher::{CacheItemRequest, CachePath};
 use crate::services::download::{DownloadStatus, RemoteDif};
 use crate::types::{ObjectId, Scope};
@@ -156,15 +157,20 @@ impl CacheItemRequest for FetchFileDataRequest {
         let future = async move {
             let status = downloader
                 .download(file_id, download_file.path().to_owned())
-                .await
-                .map_err(Self::Error::from)?;
+                .await;
 
             match status {
-                DownloadStatus::NotFound => {
+                Ok(DownloadStatus::NotFound) => {
                     log::debug!("No debug file found for {}", cache_key);
                     return Ok(CacheStatus::Negative);
                 }
-                DownloadStatus::Completed => {
+
+                Err(e) => {
+                    log::error!("Error while downloading file: {}", LogError(&e));
+                    return Ok(CacheStatus::Negative);
+                }
+
+                Ok(DownloadStatus::Completed) => {
                     // fall-through
                 }
             }
@@ -264,5 +270,112 @@ impl CacheItemRequest for FetchFileDataRequest {
         object_handle.configure_scope();
 
         object_handle
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+    use std::time::Duration;
+
+    use crate::cache::{Cache, CacheStatus};
+    use crate::config::{CacheConfig, CacheConfigs, Config};
+    use crate::services::download::DownloadService;
+    use crate::services::objects::data_cache::Scope;
+    use crate::services::objects::{FindObject, ObjectPurpose, ObjectsActor};
+    use crate::sources::FileType;
+    use crate::test::{self, tempdir};
+
+    use symbolic::common::DebugId;
+    use tempfile::TempDir;
+
+    fn objects_actor(tempdir: &TempDir) -> ObjectsActor {
+        let meta_cache = Cache::from_config(
+            "meta",
+            Some(tempdir.path().join("meta")),
+            None,
+            CacheConfig::from(CacheConfigs::default().derived),
+        )
+        .unwrap();
+
+        let data_cache = Cache::from_config(
+            "data",
+            Some(tempdir.path().join("data")),
+            None,
+            CacheConfig::from(CacheConfigs::default().downloaded),
+        )
+        .unwrap();
+
+        let config = Arc::new(Config {
+            connect_to_reserved_ips: true,
+            download_timeout: Duration::from_millis(10),
+            ..Config::default()
+        });
+
+        let download_svc = DownloadService::new(config);
+        ObjectsActor::new(meta_cache, data_cache, download_svc)
+    }
+
+    #[tokio::test]
+    async fn test_negative_cache() {
+        test::setup();
+
+        let server = test::FailingSymbolServer::new();
+        let cachedir = tempdir();
+        let objects_actor = objects_actor(&cachedir);
+
+        test::spawn_compat(move || async move {
+            let find_object = FindObject {
+                // A request for a bcsymbolmap will expand to only one file that is being looked up.
+                // Other filetypes will lead to multiple requests, trying different file extensions, etc
+                filetypes: &[FileType::BcSymbolMap],
+                purpose: ObjectPurpose::Debug,
+                scope: Scope::Global,
+                identifier: DebugId::default().into(),
+                sources: Arc::new([]),
+            };
+
+            // for each of the different symbol sources, we assert that:
+            // * we get a negative cache result no matter how often we try
+            // * we hit the symbol source exactly once for the initial request
+            // * the second try should *not* hit the symbol source, but should rather be served by the cache
+
+            // server rejects the request (500)
+            let find_object = FindObject {
+                sources: Arc::new([server.reject_source.clone()]),
+                ..find_object
+            };
+            let result = objects_actor.find(find_object.clone()).await.unwrap();
+            assert_eq!(result.meta.unwrap().status, CacheStatus::Negative);
+            assert_eq!(server.accesses(), 1);
+            let result = objects_actor.find(find_object.clone()).await.unwrap();
+            assert_eq!(result.meta.unwrap().status, CacheStatus::Negative);
+            assert_eq!(server.accesses(), 0);
+
+            // server responds with not found (404)
+            let find_object = FindObject {
+                sources: Arc::new([server.not_found_source.clone()]),
+                ..find_object
+            };
+            let result = objects_actor.find(find_object.clone()).await.unwrap();
+            assert_eq!(result.meta.unwrap().status, CacheStatus::Negative);
+            assert_eq!(server.accesses(), 1);
+            let result = objects_actor.find(find_object.clone()).await.unwrap();
+            assert_eq!(result.meta.unwrap().status, CacheStatus::Negative);
+            assert_eq!(server.accesses(), 0);
+
+            // server accepts the request, but never sends any reply (timeout)
+            let find_object = FindObject {
+                sources: Arc::new([server.pending_source.clone()]),
+                ..find_object
+            };
+            let result = objects_actor.find(find_object.clone()).await.unwrap();
+            assert_eq!(result.meta.unwrap().status, CacheStatus::Negative);
+            assert_eq!(server.accesses(), 1);
+            let result = objects_actor.find(find_object.clone()).await.unwrap();
+            assert_eq!(result.meta.unwrap().status, CacheStatus::Negative);
+            assert_eq!(server.accesses(), 0);
+        })
+        .await;
     }
 }