+++ crates/symbolicator/src/test.rs
@@ -19,11 +19,14 @@
 use std::future::Future;
 use std::net::SocketAddr;
 use std::path::{Path, PathBuf};
+use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 
 use futures::{FutureExt, TryFutureExt};
 use log::LevelFilter;
 use reqwest::Url;
+use warp::filters::fs::File;
+use warp::reject::{Reject, Rejection};
 use warp::Filter;
 
 use crate::sources::{
@@ -239,5 +242,89 @@ pub(crate) fn symbol_server() -> (Server, SourceConfig) {
     (server, source)
 }
 
+#[derive(Debug, Clone, Copy)]
+struct GoAway;
+
+impl Reject for GoAway {}
+
+pub(crate) struct FailingSymbolServer {
+    #[allow(unused)]
+    server: Server,
+    times_accessed: Arc<AtomicUsize>,
+    pub(crate) reject_source: SourceConfig,
+    pub(crate) pending_source: SourceConfig,
+    pub(crate) not_found_source: SourceConfig,
+}
+
+impl FailingSymbolServer {
+    pub(crate) fn new() -> Self {
+        let times_accessed = Arc::new(AtomicUsize::new(0));
+
+        let times = times_accessed.clone();
+        let reject = warp::path("reject").and_then(move || {
+            let times = Arc::clone(&times);
+            async move {
+                (*times).fetch_add(1, Ordering::SeqCst);
+
+                Err::<File, _>(warp::reject::custom(GoAway))
+            }
+        });
+
+        let times = times_accessed.clone();
+        let not_found = warp::path("not-found").and_then(move || {
+            let times = Arc::clone(&times);
+            async move {
+                (*times).fetch_add(1, Ordering::SeqCst);
+
+                Err::<File, _>(warp::reject::not_found())
+            }
+        });
+
+        let times = times_accessed.clone();
+        let pending = warp::path("pending").and_then(move || {
+            (*times).fetch_add(1, Ordering::SeqCst);
+
+            std::future::pending::<Result<File, Rejection>>()
+        });
+
+        let server = Server::new(reject.or(not_found).or(pending));
+
+        let files_config =
+            CommonSourceConfig::with_layout(crate::sources::DirectoryLayoutType::Unified);
+
+        let reject_source = SourceConfig::Http(Arc::new(HttpSourceConfig {
+            id: SourceId::new("reject"),
+            url: server.url("reject/"),
+            headers: Default::default(),
+            files: files_config.clone(),
+        }));
+
+        let pending_source = SourceConfig::Http(Arc::new(HttpSourceConfig {
+            id: SourceId::new("pending"),
+            url: server.url("pending/"),
+            headers: Default::default(),
+            files: files_config.clone(),
+        }));
+
+        let not_found_source = SourceConfig::Http(Arc::new(HttpSourceConfig {
+            id: SourceId::new("not-found"),
+            url: server.url("not-found/"),
+            headers: Default::default(),
+            files: files_config,
+        }));
+
+        FailingSymbolServer {
+            server,
+            times_accessed,
+            reject_source,
+            pending_source,
+            not_found_source,
+        }
+    }
+
+    pub(crate) fn accesses(&self) -> usize {
+        self.times_accessed.swap(0, Ordering::SeqCst)
+    }
+}
 // make sure procspawn works.
 procspawn::enable_test_support!();